requirements.txt
asgiref==3.7.2
attrs==22.2.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
colorama==0.4.6
coverage==7.0.5
crashtest==0.4.1
distlib==0.3.7
Django==5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
dulwich==0.21.6
fastjsonschema==2.19.0
filelock==3.13.1
idna==3.6
importlib-metadata==6.8.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.3.0
keyring==24.3.0
more-itertools==10.1.0
msgpack==1.0.7
packaging==23.0
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.0.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
psycopg2==2.9.9
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.9.0
pyproject_hooks==1.0.0
pytest==7.2.1
pytest-cov==4.0.0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.2
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
shellingham==1.5.4
simplejwt==2.0.1
sqlparse==0.4.4
tomlkit==0.12.3
trove-classifiers==2023.11.22
typing==3.7.4.3
uritemplate==4.1.1
urllib3==2.1.0
virtualenv==20.24.7
xattr==0.10.1
zipp==3.17.0
